feat(build-configuration-dsl): Improves build configuration dsl
