+++ functional-tests/test-bed/src/main/java/org/arquillian/smart/testing/ftest/testbed/project/ProjectBuilder.java
@@ -6,10 +6,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Properties;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.arquillian.smart.testing.ftest.testbed.testresults.TestResult;
@@ -17,8 +14,6 @@ import org.jboss.shrinkwrap.resolver.api.maven.embedded.BuiltProject;
 import org.jboss.shrinkwrap.resolver.api.maven.embedded.EmbeddedMaven;
 import org.jboss.shrinkwrap.resolver.api.maven.embedded.pom.equipped.PomEquippedEmbeddedMaven;
 
-import static java.lang.System.getProperty;
-import static java.util.Arrays.stream;
 import static org.arquillian.smart.testing.ftest.testbed.testresults.Status.FAILURE;
 import static org.arquillian.smart.testing.ftest.testbed.testresults.Status.PASSED;
 import static org.arquillian.smart.testing.ftest.testbed.testresults.SurefireReportReader.loadTestResults;
@@ -26,129 +21,24 @@ import static org.arquillian.smart.testing.ftest.testbed.testresults.SurefireRep
 public class ProjectBuilder {
 
     private static final String TEST_REPORT_PREFIX = "TEST-";
-    /*
-    MAVEN_DEBUG_OPTS="-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=8000"
-     */
-    private static final String MVN_DEBUG_AGENT = "-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=%s,address=%s";
-    private static final String SUREFIRE_DEBUG_SETTINGS = " -Xnoagent -Djava.compiler=NONE";
-    private static final int DEFAULT_DEBUG_PORT = 8000;
-    private static final int DEFAULT_SUREFIRE_DEBUG_PORT = 5005;
 
     private final Path root;
-    private final Project project;
-    private final Properties systemProperties = new Properties();
-    private final Set<String> excludedProjects = new HashSet<>();
+    private final BuildConfigurator buildConfigurator;
 
-    private int remotePort = DEFAULT_DEBUG_PORT;
-    private int surefireRemotePort = DEFAULT_SUREFIRE_DEBUG_PORT;
-    private boolean quietMode = true;
-    private boolean enableRemoteDebugging = false;
-    private boolean suspend = true;
-    private boolean mvnDebugOutput;
-    private boolean enableSurefireRemoteDebugging = false;
-    private String mavenOpts = "-Xms512m -Xmx1024m";
-
-    ProjectBuilder(Path root, Project project) {
-        systemProperties.put("surefire.exitTimeout", "-1"); // see http://bit.ly/2vARQ5p
-        systemProperties.put("surefire.timeout", "0"); // see http://bit.ly/2u7xCAH
+    ProjectBuilder(Path root) {
         this.root = root;
-        this.project = project;
+        this.buildConfigurator = new BuildConfigurator(this);
     }
 
-    public Project configure() {
-        return this.project;
-    }
-
-    /**
-     * Enables remote debugging of embedded maven build so we can troubleshoot our extension and provider
-     * By default it sets suspend to 'y', so the build will wait until we attach remote debugger to the port DEFAULT_DEBUG_PORT
-     */
-    public ProjectBuilder withRemoteDebugging() {
-        return withRemoteDebugging(DEFAULT_DEBUG_PORT, true);
-    }
-
-    public ProjectBuilder withRemoteDebugging(int port) {
-        return withRemoteDebugging(port, true);
-    }
-
-    public ProjectBuilder withRemoteDebugging(int port, boolean suspend) {
-        this.enableRemoteDebugging = true;
-        this.remotePort = port;
-        this.suspend = suspend;
-        return this;
-    }
-
-    public ProjectBuilder withRemoteSurefireDebugging() {
-        return withRemoteSurefireDebugging(DEFAULT_SUREFIRE_DEBUG_PORT, true);
-    }
-
-    public ProjectBuilder withRemoteSurefireDebugging(int surefireRemotePort) {
-        return withRemoteSurefireDebugging(surefireRemotePort, true);
-    }
-
-    public ProjectBuilder withRemoteSurefireDebugging(int surefireRemotePort, boolean suspend) {
-        this.surefireRemotePort = surefireRemotePort;
-        this.suspend = suspend;
-        this.enableSurefireRemoteDebugging = true;
-        return this;
-    }
-
-    public ProjectBuilder quiet(boolean quiet) {
-        this.quietMode = quiet;
-        return this;
-    }
-
-    public ProjectBuilder logBuildOutput() {
-        return logBuildOutput(false);
-    }
-
-    private ProjectBuilder logBuildOutput(boolean debug) {
-        withDebugOutput(debug);
-        return quiet(false);
-    }
-
-    private ProjectBuilder withDebugOutput(boolean debug) {
-        this.mvnDebugOutput = debug;
-        quiet(!debug);
-        return this;
-    }
-
-    /**
-     * Enables mvn debug output (-X) flag. Implies build logging output.
-     */
-    public ProjectBuilder withDebugOutput() {
-        return withDebugOutput(true);
-    }
-
-    public ProjectBuilder withSystemProperties(String ... systemPropertiesPairs) {
-        if (systemPropertiesPairs.length % 2 != 0) {
-            throw new IllegalArgumentException("Expecting even amount of variable name - value pairs to be passed. Got "
-                + systemPropertiesPairs.length
-                + " entries. "
-                + Arrays.toString(systemPropertiesPairs));
-        }
-
-        for (int i = 0; i < systemPropertiesPairs.length; i += 2) {
-            this.systemProperties.put(systemPropertiesPairs[i], systemPropertiesPairs[i + 1]);
-        }
-
-        return this;
-    }
-
-    public ProjectBuilder excludeProjects(String... projects) {
-        this.excludedProjects.addAll(
-            stream(projects)
-                .map(excludeProject -> "!" + excludeProject)
-                .collect(Collectors.toList())
-        );
-        return this;
+    public BuildConfigurator options() {
+        return this.buildConfigurator;
     }
 
     List<TestResult> build(String... goals) {
         final PomEquippedEmbeddedMaven embeddedMaven =
             EmbeddedMaven.forProject(root.toAbsolutePath().toString() + "/pom.xml");
 
-        enableDebugOptions(embeddedMaven);
+        buildConfigurator.enableDebugOptions(embeddedMaven);
 
         final String mvnInstallation = System.getenv("TEST_BED_M2_HOME");
         if (mvnInstallation != null) {
@@ -158,12 +48,12 @@ public class ProjectBuilder {
         final BuiltProject build = embeddedMaven
                     .setShowVersion(true)
                     .setGoals(goals)
-                    .setProjects(excludedProjects.toArray(new String[excludedProjects.size()]))
-                    .setDebug(isMavenDebugOutputEnabled())
-                    .setQuiet(disableQuietWhenAnyDebugModeEnabled() && quietMode)
+                    .setProjects(buildConfigurator.getExcludedProjects())
+                    .setDebug(buildConfigurator.isMavenDebugOutputEnabled())
+                    .setQuiet(buildConfigurator.disableQuietWhenAnyDebugModeEnabled() && buildConfigurator.isQuietMode())
                     .skipTests(false)
-                    .setProperties(systemProperties)
-                    .setMavenOpts(mavenOpts)
+                    .setProperties(buildConfigurator.getSystemProperties())
+                    .setMavenOpts(buildConfigurator.getMavenOpts())
                     .ignoreFailure()
                 .build();
 
@@ -173,26 +63,6 @@ public class ProjectBuilder {
         }
 
         return accumulatedTestResults();
-    }
-
-    private boolean disableQuietWhenAnyDebugModeEnabled() {
-        return !isMavenDebugOutputEnabled() && !isSurefireRemoteDebuggingEnabled() && !isRemoteDebugEnabled();
-    }
-
-    private void enableDebugOptions(PomEquippedEmbeddedMaven embeddedMaven) {
-        if (isRemoteDebugEnabled()) {
-            final String debugOptions = String.format(MVN_DEBUG_AGENT, shouldSuspend(), getRemotePort());
-            addMavenOpts(debugOptions);
-        }
-
-        if (isSurefireRemoteDebuggingEnabled()) {
-            this.systemProperties.put("maven.surefire.debug",
-                String.format(MVN_DEBUG_AGENT, shouldSuspend(), getSurefireDebugPort()) + SUREFIRE_DEBUG_SETTINGS);
-        }
-    }
-
-    private void addMavenOpts(String options) {
-        this.mavenOpts += " " + options;
     }
 
     private List<TestResult> accumulatedTestResults() {
@@ -218,30 +88,18 @@ public class ProjectBuilder {
         }
     }
 
-    private boolean isRemoteDebugEnabled() {
-        return Boolean.valueOf(getProperty("test.bed.mvn.remote.debug", Boolean.toString(this.enableRemoteDebugging)));
+    public List<TestResult> run() {
+        return run("clean", "package");
     }
 
-    boolean isSurefireRemoteDebuggingEnabled() {
-        return Boolean.valueOf(getProperty("test.bed.mvn.surefire.remote.debug", Boolean.toString(this.enableSurefireRemoteDebugging)));
+    public List<TestResult> run(boolean log) {
+        if (log) {
+            buildConfigurator.logBuildOutput();
+        }
+        return run("clean", "package");
     }
 
-    private String shouldSuspend() {
-        final Boolean suspend =
-            Boolean.valueOf(getProperty("test.bed.mvn.remote.debug.suspend", Boolean.toString(this.suspend)));
-        return suspend ? "y" : "n";
-    }
-
-    int getRemotePort() {
-        return Integer.valueOf(getProperty("test.bed.mvn.remote.debug.port", Integer.toString(this.remotePort)));
-    }
-
-    boolean isMavenDebugOutputEnabled() {
-        return Boolean.valueOf(getProperty("test.bed.mvn.debug.output", Boolean.toString(this.mvnDebugOutput)));
-    }
-
-    int getSurefireDebugPort() {
-        return Integer.valueOf(
-            getProperty("test.bed.mvn.surefire.remote.debug.port", Integer.toString(this.surefireRemotePort)));
+    public List<TestResult> run(String... goals) {
+        return build(goals);
     }
 }
